refactor(api-headless-cms): change structure of es helper functions
