+++ packages/api-headless-cms/src/content/plugins/crud/contentEntry/es/createElasticsearchLimit.ts
@@ -1,0 +1,15 @@
+const ES_LIMIT_MAX = 10000;
+const ES_LIMIT_DEFAULT = 50;
+
+export const createElasticsearchLimit = (
+    limit: number,
+    defaultValue = ES_LIMIT_DEFAULT
+): number => {
+    if (!limit) {
+        return defaultValue;
+    }
+    if (limit < ES_LIMIT_MAX) {
+        return limit;
+    }
+    return ES_LIMIT_MAX - 1;
+};
