+++ packages/api-headless-cms/src/content/plugins/crud/contentEntry/es/index.ts
@@ -1,0 +1,3 @@
+export * from "./createElasticsearchLimit";
+export * from "./createElasticsearchQueryBody";
+export * from "./entryIndexHelpers";
